Refactor: UX 개선
- 화면 및 경로 전환시 슬라이드 트랜지션
- 우측 하단 플로팅 버튼 위치 변경시 트랜지션
- '이름바꾸기' 메뉴 선택시 자동 포커스 그리고 포커스 벗어날 시 자동저장
- '이름바꾸기' 동일명칭시 덮어쓰는 버그 수정
- '최근 작성 목록' 카드 컴포넌트화 및 스타일시트 수정

+++ src/tools/fs.ts
@@ -163,6 +163,9 @@ const rename = async (
     rename = `${rename}${ext ? ext : ''}`
     // 새로운 경로를 생성
     const renameDir = path.join(dir, rename)
+    if (fs.existsSync(renameDir)) {
+        throw new Error('이미 존재하는 파일명 입니다.')
+    }
     fs.renameSync(tarDir, renameDir)
     return sliceRootDir(rootDir, renameDir)
 }
